re-run setup.py and add function for ga api caller to deal with next page token
