+++ database_connection/__init__.py
@@ -1,3 +1,3 @@
 from .sql_connection import get_connection, run_sql, run_sql_text_query, row_to_df, \
     drop_table, delete_records, turn_data_into_insert
-from .google_api_query import ga_api_caller, pivot_report_df, metric_report_df, get_dimension_row_names, run_report
+from .google_analytics_api_query import ga_api_caller, pivot_report_df, metric_report_df, get_dimension_row_names, create_service
